### 2018-06-15 #777 @whedon
[PRE REVIEW]: quimb: a python library for quantum information and many-body calculations
**Submitting author:** @jcmgray (<a href="http://orcid.org/0000-0001-9461-3024">Johnnie Gray</a>)
**Repository:** <a href="https://github.com/jcmgray/quimb" target ="_blank">https://github.com/jcmgray/quimb</a>
**Version:** 0.4.2
