### 2018-06-16 #777 @jedbrown
@mlxd :wave: would you be available to review?
