### 2018-06-18 103e1cf
improvements to NNI ham generation and docs
