+++ quimb/tensor/tensor_gen.py
@@ -1,13 +1,12 @@
 """Generate specific tensor network states and operators.
 """
-import functools
 import random
 
 import numpy as np
 
 from ..accel import make_immutable
 from ..linalg.base_linalg import norm_fro_dense
-from ..gen.operators import spin_operator, eye
+from ..gen.operators import spin_operator, eye, _gen_mbl_random_factors
 from ..gen.rand import randn
 from .tensor_core import Tensor
 from .tensor_1d import MatrixProductState, MatrixProductOperator
@@ -355,7 +354,8 @@ def spin_ham_mpo_tensor(one_site_terms, two_site_terms, S=1 / 2,
 
     Returns
     -------
-    numpy.ndarray{, numpy.ndarray, numpy.ndarray}
+    numpy.ndarray[, numpy.ndarray, numpy.ndarray]
+        The middle, left, right or all three MPO tensors.
     """
     # local dimension
     D = int(2 * S + 1)
@@ -407,9 +407,9 @@ def spin_ham_mpo_tensor(one_site_terms, two_site_terms, S=1 / 2,
 
 
 class SpinHam:
-    """Class for easily building translationally invariant spin hamiltonians in
-    MPO or NNI form. Currently limited to nearest neighbour interactions (and
-    single site terms).
+    """Class for easily building custom spin hamiltonians in MPO or NNI form.
+    Currently limited to nearest neighbour interactions (and single site
+    terms).
 
     Parameters
     ----------
@@ -418,16 +418,32 @@ class SpinHam:
     cyclic : bool, optional
         Whether to use periodic boundary conditions - default is False.
 
-    Example
-    -------
-    >>> builder = SpinHam(S=3 / 2)
-    >>> builder.add_term(-0.3, 'Z')
-    >>> builder.add_term(0.5, '+', '-')
-    >>> builder.add_term(0.5, '-', '+')
-    >>> builder.add_term(1.0, 'Z', 'Z')
-    >>> mpo_ham = builder.build_mpo(100)
-    >>> mpo_ham
-    <MatrixProductOperator(tensors=100, structure='I{}', nsites=100)>
+    Examples
+    --------
+    Initialize the spin hamiltonian builder:
+
+        >>> builder = SpinHam(S=3 / 2)
+
+    Add some two-site terms:
+
+        >>> builder += 0.5, '+', '-'
+        >>> builder += 0.5, '-', '+'
+        >>> builder += 1.0, 'Z', 'Z'
+
+    Add a single site term:
+
+        >>> builder -= 0.3, 'Z'
+
+    Build a MPO version of the hamiltonian for use with DMRG:
+
+        >>> mpo_ham = builder.build_mpo(100)
+        >>> mpo_ham
+        <MatrixProductOperator(tensors=100, structure='I{}', nsites=100)>
+
+    Build a NNI version of the hamiltonian for use with TEBD:
+
+        >>> builder.build_nni(100)
+        <NNI(n=100, cyclic=False)>
     """
 
     def __init__(self, S=1 / 2, cyclic=False):
@@ -456,6 +472,19 @@ class SpinHam:
         else:
             raise NotImplementedError("3-body+ terms are not supported yet.")
 
+    def sub_term(self, factor, *operators):
+        """Subtract a term - simple alias that flips sign of ``factor``.
+        """
+        self.add_term(-factor, *operators)
+
+    def __iadd__(self, term):
[…]
