### 2018-06-18 3f77c7e
doc updates
