+++ docs/develop.rst
@@ -1,25 +1,27 @@
 ###############
-Developer notes
+Developer Notes
 ###############
+
+
+Contributing
+============
+
+Things to check if new functionality added:
+
+1. Ensure functions are unit tested.
+2. Ensure functions have `numpy style docstrings <http://sphinxcontrib-napoleon.readthedocs.io/en/latest/example_numpy.html>`_.
+3. Ensure code is PEP8 compliant.
+4. If module, add to ``docs/api.rst`` for autosummarizing.
+5. Add to ``quimb/__init__.py`` and ``"__all__"`` if appropriate (or the
+   tensor network equivalent ``quimb.tensor.__init__.py``).
 
 
 Running the Tests
 =================
 
-Testing ``quimb`` requires `pytest <https://docs.pytest.org/en/latest/index.html>`_ and simply involves running ``pytest`` in the root ``quimb`` directory.
+Testing ``quimb`` requires `pytest <https://docs.pytest.org/en/latest/index.html>`_ (as well as ``coverage`` and ``pytest-cov``) and simply involves running ``pytest`` in the root ``quimb`` directory.
 
 The tests can also be run with pre-spawned mpi workers using the command ``quimb-mpi-python -m pytest`` (but not in syncro mode -- see :ref:`mpistuff`).
-
-
-Adding a function
-=================
-
-Steps:
-
-1. Ensure function has numpy style docstring.
-2. Make sure function is unit tested.
-3. If module, add to ``docs/api.rst`` for autosummarizing.
-4. Add to ``quimb/__init__.py`` and ``"__all__"`` if appropriate.
 
 
 Building the docs locally
@@ -29,4 +31,20 @@ Building the docs requires `sphinx <http://www.sphinx-doc.org/en/stable/>`_, `sp
 
 1. To start from scratch, remove ``quimb/docs/_autosummary`` and ``quimb/docs/_build``.
 2. Run ``make html`` (``make.bat html`` on windows) in the ``quimb/docs`` folder.
-3. Launch page: ``quimb/docs/_build/html/index.html``.
+3. Launch the page: ``quimb/docs/_build/html/index.html``.
+
+
+Minting a Release
+=================
+
+``quimb`` uses `versioneer <https://github.com/warner/python-versioneer>`_
+to manage versions and releases. The steps to release a new version
+on `pypi <https://pypi.org>`_  are as follows:
+
+1. Make sure all tests are passing, as well as the continuous integration
+   and readthedocs build.
+2. ``git tag`` the release with next ``X.Y.Z`` (n.b. no 'v' prefix).
+3. Remove any old builds: ``rm dist/*```
+4. Build the tar and wheel ``python setup.py bdist_wheel sdist``
+5. Optionally remove the ``build`` folder.
+6. Upload using twine: ``twine upload dist/*``
