+++ quimb/tensor/tensor_gen.py
@@ -376,8 +376,8 @@ def spin_ham_mpo_tensor(one_site_terms, two_site_terms, S=1 / 2,
             s1 = spin_operator(s1, S=S)
         if isinstance(s2, str):
             s2 = spin_operator(s2, S=S)
-        H[1 + i, 0, :, :] = s1
-        H[-1, 1 + i, :, :] = factor * s2
+        H[1 + i, 0, :, :] = factor * s1
+        H[-1, 1 + i, :, :] = s2
 
     H[0, 0, :, :] = eye(D)
     H[B - 1, B - 1, :, :] = eye(D)
@@ -406,10 +406,34 @@ def spin_ham_mpo_tensor(one_site_terms, two_site_terms, S=1 / 2,
         return HL, H, HR
 
 
+class _TermAdder:
+    """Simple class to allow ``SpinHam`` syntax like
+    ``builder[i, j] += (1/2, 'Z', 'X')``. This object is temporarily created
+    by the getitem call, accumulates the new term, then has its the new
+    combined list of terms extracted in the setitem call.
+    """
+
+    def __init__(self, terms, nsite):
+        self.terms = terms
+        self.nsite = nsite
+
+    def __iadd__(self, new):
+        if len(new) != self.nsite + 1:
+            raise ValueError(
+                "New terms should be of the form")
+
+        if self.terms is None:
+            self.terms = [new]
+        else:
+            self.terms += [new]
+        return self
+
+
 class SpinHam:
     """Class for easily building custom spin hamiltonians in MPO or NNI form.
     Currently limited to nearest neighbour interactions (and single site
-    terms).
+    terms). It is possible to set 'default' translationally invariant terms,
+    but also terms acting on specific sites only (which take precedence).
 
     Parameters
     ----------
@@ -444,6 +468,19 @@ class SpinHam:
 
         >>> builder.build_nni(100)
         <NNI(n=100, cyclic=False)>
+
+    You can also set terms for specific sites (this overides any of the
+    'default', translationally invariant terms set as above):
+
+        >>> builder[10, 11] += 0.75, '+', '-'
+        >>> builder[10, 11] += 0.75, '-', '+'
+        >>> builder[10, 11] += 1.5, 'Z', 'Z'
+
+    Or specific one-site terms (which again overides any default
+    single site terms set above):
+
+        >>> builder[10] += 3.7, 'Z'
+        >>> builder[11] += 0.0, 'I' # '0' term turns off field
     """
 
     def __init__(self, S=1 / 2, cyclic=False):
@@ -451,6 +488,10 @@ class SpinHam:
         self.one_site_terms = []
         self.two_site_terms = []
         self.cyclic = cyclic
+
+        # Holders for any non-translationally invariant terms
+        self.var_one_site_terms = {}
+        self.var_two_site_terms = {}
 
     def add_term(self, factor, *operators):
         """Add another term to the expression to be built.
@@ -485,21 +526,108 @@ class SpinHam:
         self.sub_term(*term)
         return self
 
+    def __getitem__(self, sites):
+        """Part of the machinery that allows terms to be added to specific
+        sites like::
+
+            >>> builder[i] += 1/2, 'X'
+            >>> builder[45, 46] += 1/2, 'Z', 'Z'
+
+        """
+        if isinstance(sites, int):
+            return _TermAdder(self.var_one_site_terms.get(sites, None), 1)
+
+        i, j = sorted(sites)
+        if j - i != 1:
+            raise ValueError("Can only add nearest neighbour terms.")
+
+        return _TermAdder(self.var_two_site_terms.get(sites, None), 2)
+
[…]
