### 2018-06-20 c2331c5
support new opt_einsum ``get_symbol``
