### 2018-06-20 fe0d1b6
TN: fixes & docs for non-trans-invar NNI and TEBD
