### 2018-06-20 c0b3484
TN: add fallback svd with alt driver
very occasionally normal SVD doesn't converge
