### 2018-06-21 bcf5d01
fixes for single precision DMRG
