### 2018-06-22 e19e3bb
various doc tweaks [ci skip]
