### 2018-06-25 ef5bf4a
approx_spectral: fix bug where huge fitted est triggered convergence
