### 2018-06-25 08fa435
add MatrixProductState.from_dense
