+++ quimb/tensor/tensor_1d.py
@@ -1021,6 +1021,93 @@ class MatrixProductState(TensorNetwork1DVector,
         super().__init__(gen_tensors(), structure=site_tag_id, sites=sites,
                          nsites=nsites, check_collisions=False, **tn_opts)
 
+    @classmethod
+    def from_TN(cls, tn, site_ind_id, site_tag_id,
+                cyclic=False, inplace=False):
+        """Convert a ``TensorNetwork`` into a ``MatrixProductState``, assuming
+        it has the appropirate underlying structure.
+
+        Parameters
+        ----------
+        tn : TensorNetwork
+            The tensor network to convert.
+        site_ind_id : str
+            The string formatter that specifies the site indices -- it should
+            match what the tensors of ``tn`` already have.
+        site_tag_id : str
+            The string formatter that specifies the site tags -- it should
+            match what the tensors of ``tn`` already have.
+        inplace : bool, optional
+            If True, perform the conversion in-place.
+        """
+        if not inplace:
+            tn = tn.copy()
+        tn.__class__ = cls
+        tn._site_ind_id = site_ind_id
+        tn._site_tag_id = site_tag_id
+        tn.cyclic = cyclic
+        return tn
+
+    @classmethod
+    def from_dense(cls, psi, dims, site_ind_id='k{}',
+                   site_tag_id='I{}', **split_opts):
+        """Create a ``MatrixProductState`` directly from a dense vector
+
+        Parameters
+        ----------
+        psi : array_like
+            The dense state to convert to MPS from.
+        dims : sequence of int
+            Physical subsystem dimensions of each site.
+        site_ind_id : str, optional
+            How to index the physical sites, see
+            :class:`~quimb.tensor.tensor_1d.MatrixProductState`.
+        site_tag_id : str, optional
+            How to tag the physical sites, see
+            :class:`~quimb.tensor.tensor_1d.MatrixProductState`.
+        split_opts
+            Supplied to :func:`~quimb.tensor.tensor_core.tensor_split` to
+            in order to partition the dense vector into tensors.
+
+        Returns
+        -------
+        MatrixProductState
+
+        Examples
+        --------
+
+            >>> dims = [2, 2, 2, 2, 2, 2]
+            >>> psi = rand_ket(prod(dims))
+            >>> mps = MatrixProductState.from_dense(psi, dims)
+            >>> mps.show()
+             2 4 8 4 2
+            o-o-o-o-o-o
+            | | | | | |
+        """
+        n = len(dims)
+        inds = [site_ind_id.format(i) for i in range(n)]
+
+        T = Tensor(psi.A.reshape(dims), inds=inds)
+
+        def gen_tensors():
+            #           split
+            #       <--  : yield
+            #            : :
+            #     OOOOOOO--O-O-O
+            #     |||||||  | | |
+            #     .......
+            #    left_inds
+            TM = T
+            for i in range(n - 1, 0, -1):
+                TM, TR = TM.split(left_inds=inds[:i], get='tensors',
+                                  rtags={site_tag_id.format(i)}, **split_opts)
+                yield TR
+            TM.tags.add(site_tag_id.format(0))
+            yield TM
+
+        tn = TensorNetwork(gen_tensors(), structure='I{}')
+        return cls.from_TN(tn, site_ind_id, site_tag_id)
+
     def imprint(self, other):
         """Cast ``other`` into a ``MatrixProductState`` like ``self``.
         """
@@ -1330,9 +1417,9 @@ class MatrixProductState(TensorNetwork1DVector,
                 rho.drop_tags(self.site_tag(i))
 
[…]
