### 2018-06-26 52e0571
TN: fuse_multibonds
