### 2018-06-26 969edc9
approx_spectral_tweaks
