+++ quimb/tensor/tensor_dmrg.py
@@ -151,26 +151,26 @@ class MovingEnvironment:
 
         /-----------------------------------------------\
         +-A-A-A-A-A-A-A-A-A-A-A-A-B-B-B-B-B-B-B-B-B-B-B-+
-          | | | | | | | | | | | | | | | | | | | | | | |           -->
+          | | | | | | | | | | | | | | | | | | | | | | |           ==>
         +-A-A-A-A-A-A-A-A-A-A-A-A-B-B-B-B-B-B-B-B-B-B-B-+
         \-----------------------------------------------/
 
         +~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~+
-        |   /-A-A-A-A-A-A-A-A-A-A-A-A-\   |                       -->
+        |   /-A-A-A-A-A-A-A-A-A-A-A-A-\   |                       ==>
         +~<BL | | | | | | | | | | | | BR>~+
             \-A-A-A-A-A-A-A-A-A-A-A-A-/
               ^                     ^
         segment_start          segment_stop - 1
 
         +~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~+
-        |   /-A-A-\                        |                      -->
+        |   /-A-A-\                        |                      ==>
         +~<BL | |  AAAAAAAAAAAAAAAAAAAABR>~+
             \-A-A-/
               ...
             <-bsz->
 
         +~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~+
-        |               /-A-A-\           |                       -->
+        |               /-A-A-\           |                       ==>
         +~<BLAAAAAAAAAAA  | |  AAAAAAABR>~+
                         \-A-A-/
                           i i+1
