+++ quimb/tensor/tensor_1d.py
@@ -425,7 +425,7 @@ class TensorNetwork1DFlat:
 
                 i                i
                -o-o-            ->-s-
-            ... | | ...  --> ... | | ...
+            ... | | ...  ==> ... | | ...
 
         Parameters
         ----------
@@ -442,7 +442,7 @@ class TensorNetwork1DFlat:
 
                   i                i
                -o-o-            -s-<-
-            ... | | ...  --> ... | | ...
+            ... | | ...  ==> ... | | ...
 
         Parameters
         ----------
@@ -460,7 +460,7 @@ class TensorNetwork1DFlat:
 
                           i              i
             >->->->->->->-o-o-         +-o-o-
-            | | | | | | | | | ...  ->  | | | ...
+            | | | | | | | | | ...  =>  | | | ...
             >->->->->->->-o-o-         +-o-o-
 
         Parameters
@@ -1207,7 +1207,7 @@ class MatrixProductState(TensorNetwork1DVector,
 
             -o-o-A-B-o-o-
              | | | | | |            -o-o-GGG-o-o-           -o-o-X~Y-o-o-
-             | | GGG | |     -->     | | | | | |     -->     | | | | | |
+             | | GGG | |     ==>     | | | | | |     ==>     | | | | | |
              | | | | | |                 i j                     i j
                  i j
 
@@ -1458,26 +1458,31 @@ class MatrixProductState(TensorNetwork1DVector,
             o-o-o-o-A-A-A-A-A-A-A-A-o-o-B-B-B-B-B-B-o-o-o-o-o-o-o-o-o
             | | | | | | | | | | | | | | | | | | | | | | | | | | | | |
 
-                                      --> form inner product
+                                      ==> form inner product
 
+                    ...............     ...........
             o-o-o-o-A-A-A-A-A-A-A-A-o-o-B-B-B-B-B-B-o-o-o-o-o-o-o-o-o
             | | | | | | | | | | | | | | | | | | | | | | | | | | | | |
             o-o-o-o-A-A-A-A-A-A-A-A-o-o-B-B-B-B-B-B-o-o-o-o-o-o-o-o-o
 
-                                      --> lateral SVD on each section
+                                      ==> lateral SVD on each section
 
                       .....sysa......     ...sysb....
                       /\             /\   /\         /\
               ... ~~~E  A~~~~~~~~~~~A  E~E  B~~~~~~~B  E~~~ ...
                       \/             \/   \/         \/
 
-                                      --> vertical SVD and unfold on A & B
+                                      ==> vertical SVD and unfold on A & B
 
                               |                 |
                       /-------A-------\   /-----B-----\
               ... ~~~E                 E~E             E~~~ ...
                       \-------A-------/   \-----B-----/
                               |                 |
+
+        With various special cases including OBC or end spins included in
+        subsytems.
+
 
         Parameters
         ----------
@@ -1590,6 +1595,12 @@ class MatrixProductState(TensorNetwork1DVector,
         compressed = []
         for section, (ul, _, ll, _) in zip(sections, ul_ur_ll_lrs):
 
+            #           section
+            #   ul -o-o-o-o-o-o-o-o-o-       ul -\       /-
+            #       | | | | | | | | |   ==>       0~~~~~0
+            #   ll -o-o-o-o-o-o-o-o-o-       ll -/   :   \-
+            #                                      hmax_bond
+
             if leave_short:
                 # if section is short doesn't make sense to lateral compress
                 #     work out roughly when this occurs by comparing bond size
@@ -1600,11 +1611,6 @@ class MatrixProductState(TensorNetwork1DVector,
                     continue
 
             section_tags = map(self.site_tag, section)
-
-            #   ul -o-o-o-o-o-o-o-o-o-       ul -\      /-
[…]
