### 2018-07-12 cdd497b
rand_rademacher and approx_linalg type support
