### 2018-07-15 jcmgray/quimb#2 @dalcinl
Missing dependency for tqdm in setup.py
It seems that `tqdm` is required to run a few tests in `tests/test_tensor/test_tensor_tebd.py`. This dependency is properly listed in `deps\requirements-py3.yml`, however it is not listed in `setup.py`. So I guess that you should either add `tqdm` in `extra_requires/test` in `setup.py`, or update the testsuite to skip tests requiring `tqdm`.

[Ref: openjournals/joss-reviews#819]
