### 2018-07-16 bb94df5
add tqdm to setup.py
resolves #2
