+++ setup.py
@@ -29,6 +29,7 @@ setup(
         'numexpr>=2.3',
         'psutil>=4.3.1',
         'cytoolz>=0.8.0',
+        'tqdm>=4',
     ],
     extras_require={
         'tests': [
