### 2018-07-16 jcmgray/quimb#2 @jcmgray
Thanks for noticing this! ``tqdm`` is a pretty light (and useful) dependency so I've just decided to add it to ``setup.py``.
