### 2018-07-16 eedec2c
enable constants, backends for tensor network linear operators
also remove auto bsz -> wasn't efficient
