+++ quimb/linalg/approx_spectral.py
@@ -29,7 +29,7 @@ from ..tensor.tensor_approx_spectral import (
 # --------------------------------------------------------------------------- #
 
 
-def lazy_ptr_linop(psi_ab, dims, sysa):
+def lazy_ptr_linop(psi_ab, dims, sysa, **linop_opts):
     r"""A linear operator representing action of partially tracing a bipartite
     state, then multiplying another 'unipartite' state::
 
@@ -71,10 +71,11 @@ def lazy_ptr_linop(psi_ab, dims, sysa):
     return (Kab & Bab).aslinearoperator(
         ['kA{}'.format(i) for i in sysa],
         ['bA{}'.format(i) for i in sysa],
+        **linop_opts
     )
 
 
-def lazy_ptr_ppt_linop(psi_abc, dims, sysa, sysb):
+def lazy_ptr_ppt_linop(psi_abc, dims, sysa, sysb, **linop_opts):
     r"""A linear operator representing action of partially tracing a tripartite
     state, partially transposing the remaining bipartite state, then
     multiplying another bipartite state::
@@ -124,6 +125,7 @@ def lazy_ptr_ppt_linop(psi_abc, dims, sysa, sysb):
     return (Kabc & Babc).aslinearoperator(
         [('bA{}' if i in sysa else 'kB{}').format(i) for i in sys_ab],
         [('kA{}' if i in sysa else 'bB{}').format(i) for i in sys_ab],
+        **linop_opts
     )
 
 
@@ -372,7 +374,7 @@ def exp_approach(x, a, b, c):
     return a * exp(x / b) + c
 
 
-def calc_est_fit(estimates, conv_n):
+def calc_est_fit(estimates, conv_n, tau):
     """Make estimate by fitting exponential convergence to estimates.
     """
     n = len(estimates)
@@ -382,7 +384,7 @@ def calc_est_fit(estimates, conv_n):
 
     try:
         # initial guess for offset, decay length and equilibrium
-        p0 = (0, n, estimates[-1])
+        p0 = (0, n, (estimates[-2] + estimates[-1]) / 2)
         ks = np.arange(len(estimates))
 
         # weight later estimates with less error as well
@@ -391,7 +393,8 @@ def calc_est_fit(estimates, conv_n):
         with warnings.catch_warnings():
             warnings.simplefilter("ignore")
             popt, pcov = curve_fit(exp_approach, ks, estimates,
-                                   p0=p0, sigma=sigma)
+                                   p0=p0, sigma=sigma, method='lm',
+                                   gtol=tau / 10, ftol=tau / 10, xtol=tau / 10)
 
         est = popt[-1]
         err = abs(pcov[-1, -1])**0.5
@@ -475,7 +478,7 @@ def single_random_estimate(A, K, bsz, beta_tol, v0, f, pos, tau, tol_scale,
         win_est, win_err = calc_est_window(estimates, mean_ests, conv_n)
 
         # try and compute an estimate and error using exponential fit
-        fit_est, fit_err = calc_est_fit(mean_ests, conv_n)
+        fit_est, fit_err = calc_est_fit(mean_ests, conv_n, tau)
 
         # take whichever has lowest error
         est, err = min((win_est, win_err), (fit_est, fit_err),
@@ -703,18 +706,7 @@ def tr_xlogx_approx(*args, **kwargs):
 #                             Specific quantities                             #
 # --------------------------------------------------------------------------- #
 
-
-def choose_bsz_from_dims(dims, subsys):
[…]
