### 2018-07-17 e0928f4
allow opt_einsum without constants
