### 2018-07-17 6b57cb6
fix perf bug where if any dim=1 opt_einsum doesn't use blas
