### 2018-07-17 25dee99
Revert "fix perf bug where if any dim=1 opt_einsum doesn't use blas"
This reverts commit 6b57cb677a487a97200ecbf98fe0ea3d729ae7fa.
