### 2018-07-17 76133b5
further fix for opt_einsum without constants support
