### 2018-07-19 8bd9dba
tweaks to approx_spectral estimate guessing with fit
