### 2018-07-19 75f7fa0
approx_spectral: automatically use single precision and cupy
