### 2018-07-20 6d5863d
only maybe convert tensor data to numpy
