+++ docs/index.rst
@@ -77,6 +77,7 @@ The following examples, generated from the notebooks in ``docs/examples``, demon
   ./examples/ex_tn_rand_uni_gate_graphs
   ./examples/ex_dmrg_periodic
   ./examples/ex_TEBD_evo
+  ./examples/ex_quantum_circuit
 
 
 Indices and tables
