### 2018-07-20 9a27905
loosen one more approx_spectral test for polyfit changes
