### 2018-07-28 9966881
TN: add function 'bonds_size'
finds the shared bonds size between tensors (networks)
