+++ quimb/tensor/__init__.py
@@ -9,6 +9,7 @@ from .tensor_core import (
     tensor_split,
     tensor_direct_product,
     bonds,
+    bonds_size,
     Tensor,
     TensorNetwork,
     TNLinearOperator1D,
@@ -69,6 +70,7 @@ __all__ = (
     "tensor_split",
     "tensor_direct_product",
     "bonds",
+    "bonds_size",
     "Tensor",
     "TensorNetwork",
     "TNLinearOperator1D",
