### 2018-07-28 10e6b3a
MPS: simplify site decomposition code
