### 2018-07-29 b9e8b50
allow lazily constructed vectors in slepc routines
