### 2018-07-30 29c8a2a
add gates: T, S, cX, cY, cZ for convenience
