+++ quimb/__init__.py
@@ -4,118 +4,415 @@ Quantum Information for Many-Body calculations.
 
 # Accelerated basic linalg operations
 from .accel import (
-    prod, isket, isbra, isop, isvec, issparse, isreal, isherm, ispos, mul, dot,
-    vdot, rdot, ldmul, rdmul, outer, explt, get_thread_pool,
+    prod,
+    isket,
+    isbra,
+    isop,
+    isvec,
+    issparse,
+    isreal,
+    isherm,
+    ispos,
+    mul,
+    dot,
+    vdot,
+    rdot,
+    ldmul,
+    rdmul,
+    outer,
+    explt,
+    get_thread_pool,
 )
 
 # Core functions
 from .core import (
-    normalize, chop, quimbify, qu, ket, bra, dop, sparse, infer_size, trace,
-    identity, eye, speye, dim_map, dim_compress, kron, kronpow, ikron,
-    pkron, permute, itrace, partial_trace, expectation, expec, nmlz, tr, ptr,
+    normalize,
+    chop,
+    quimbify,
+    qu,
+    ket,
+    bra,
+    dop,
+    sparse,
+    infer_size,
+    trace,
+    identity,
+    eye,
+    speye,
+    dim_map,
+    dim_compress,
+    kron,
+    kronpow,
+    ikron,
+    pkron,
+    permute,
+    itrace,
+    partial_trace,
+    expectation,
+    expec,
+    nmlz,
+    tr,
+    ptr,
 )
 
 # Linear algebra functions
 from .linalg.base_linalg import (
-    eigensystem, eig, eigh, eigvals, eigvalsh,
-    eigvecs, eigvecsh, eigensystem_partial, groundstate, groundenergy,
-    bound_spectrum, eigh_window, eigvalsh_window, eigvecsh_window,
-    svd, svds, norm, expm, sqrtm, expm_multiply, Lazy,
+    eigensystem,
+    eig,
+    eigh,
+    eigvals,
+    eigvalsh,
+    eigvecs,
+    eigvecsh,
+    eigensystem_partial,
+    groundstate,
+    groundenergy,
+    bound_spectrum,
+    eigh_window,
+    eigvalsh_window,
+    eigvecsh_window,
+    svd,
+    svds,
+    norm,
+    expm,
+    sqrtm,
+    expm_multiply,
+    Lazy,
 )
 from .linalg.mpi_launcher import get_mpi_pool
 
 # Generating objects
 from .gen.operators import (
-    spin_operator, pauli, hadamard, phase_gate, Rx, Ry, Rz, swap, controlled,
-    ham_heis, ham_ising, ham_XY, ham_XXZ, ham_j1j2, ham_mbl, ham_heis_2D,
-    zspin_projector, create, destroy, num, ham_hubbard_hardcore,
+    spin_operator,
+    pauli,
+    hadamard,
+    phase_gate,
+    S_gate,
+    T_gate,
+    rotation,
+    Rx,
+    Ry,
+    Rz,
+    swap,
+    controlled,
+    CNOT,
+    cX,
+    cY,
+    cZ,
+    ham_heis,
+    ham_ising,
+    ham_XY,
+    ham_XXZ,
+    ham_j1j2,
+    ham_mbl,
+    ham_heis_2D,
+    zspin_projector,
+    create,
+    destroy,
+    num,
+    ham_hubbard_hardcore,
 )
 from .gen.states import (
-    basis_vec, up, zplus, down, zminus, plus, xplus, minus, xminus, yplus,
-    yminus, bloch_state, bell_state, singlet, thermal_state, neel_state,
-    singlet_pairs, werner_state, ghz_state, w_state, levi_civita, perm_state,
[…]
