### 2018-08-01 7baa1bc
TN tweaks
