### 2018-08-01 d4847e5
TN: add contraction_complexity
