### 2018-08-08 89486ee
TN applying gates by default only propagates site tags
