### 2018-08-08 383329c
misc fixes
