### 2018-08-08 02dcf6e
add Dense1D tensor network
MPS-like TN but really has a dense full hilbert space representation under the hood in the form of a single tensor
