### 2018-08-09 72b7247
fixes for eigh_window with Lazy ops
