+++ quimb/__init__.py
@@ -10,6 +10,7 @@ from .accel import (
     isop,
     isvec,
     issparse,
+    isdense,
     isreal,
     isherm,
     ispos,
@@ -227,6 +228,7 @@ __all__ = [
     'isop',
     'isvec',
     'issparse',
+    'isdense',
     'isreal',
     'isherm',
     'ispos',
