### 2018-08-14 d82556c
change `run` argument to `seed` for clarity
