### 2018-08-14 ed95c49
add verbosity option to MPS.partial_trace_compress
