### 2018-08-15 e95b244
super fast multi-threaded random numbers with randomgen
