+++ quimb/gen/rand.py
@@ -1,63 +1,182 @@
 """Functions for generating random quantum objects and states.
 """
-from functools import reduce, wraps
+from functools import reduce, wraps, lru_cache
 import numpy as np
 import scipy.sparse as sp
+import math
+from importlib.util import find_spec
+import os
 
-from ..accel import rdmul, dot, matrixify
-from ..core import qu, ptr, kron, nmlz
-
-
-def random_seed_fn(fn):
-    """Modify ``fn`` to take a ``seed`` argument with which call
-    :func:`numpy.random.seed`.
-    """
-
-    @wraps(fn)
-    def wrapped_fn(*args, seed=None, **kwargs):
-        if seed is not None:
-            np.random.seed(seed)
-        return fn(*args, **kwargs)
-
-    return wrapped_fn
+from ..accel import rdmul, dot, matrixify, _NUM_THREAD_WORKERS, get_thread_pool
+from ..core import qu, ptr, kron, nmlz, prod
 
 
-def randn(shape, loc=0.0, scale=1.0, dtype=float):
-    """Generate normally distributed random array of certain shape and type.
-    Like :func:`numpy.random.randn` but can specify ``dtype``.
+if (
+    find_spec('randomgen') and
+    os.environ.get('QUIMB_USE_RANDOMGEN', '').lower() not in {'false', 'off'}
+):
 
-    Parameters
-    ----------
-    shape : tuple[int]
-        The shape of the array.
-    dtype : {float, complex, ...}, optional
-        The numpy data type.
+    _RANDOM_GENS = []
 
-    Returns
-    -------
-    A : array
-    """
-    # real datatypes
-    if np.issubdtype(dtype, np.floating):
-        x = np.random.normal(loc=loc, scale=scale, size=shape)
+    @lru_cache(2)
+    def _get_randomgens(num_threads):
+        """Cached generation of random number generators, allows
+        ``random_seed_fn`` and greater efficiency.
+        """
+        global _RANDOM_GENS
 
-        need2convert = dtype not in (float, np.float_)
+        num_gens = len(_RANDOM_GENS)
+        if num_gens < num_threads:
+            from randomgen import Xoroshiro128
 
-    # complex datatypes
-    elif np.issubdtype(dtype, np.complexfloating):
-        x = (np.random.normal(loc=loc, scale=scale, size=shape) +
-             1.0j * np.random.normal(loc=loc, scale=scale, size=shape))
+            # add more generators if not enough
+            for _ in range(num_threads - num_gens):
+                _RANDOM_GENS.append(Xoroshiro128())
 
-        need2convert = dtype not in (complex, np.complex_)
+        return _RANDOM_GENS[:num_threads]
 
-    else:
-        raise TypeError("dtype {} not understood - should be float or complex."
-                        "".format(dtype))
+    def random_seed_fn(fn):
+        """Modify ``fn`` to take a ``seed`` argument with which to call
+        the seeder of ``randomgen``.
+        """
 
-    if need2convert:
-        x = x.astype(dtype)
+        @wraps(fn)
+        def wrapped_fn(*args, seed=None, **kwargs):
+            if seed is not None:
+                rg, = _get_randomgens(1)
+                rg.seed(seed)
+            return fn(*args, **kwargs)
 
-    return x
+        return wrapped_fn
+
+    @random_seed_fn
+    def randn(shape, dtype=complex, num_threads=None, scale=1.0, loc=0.0):
+        """Fast multithreaded generation of random normally distributed data
+        using ``randomgen``.
+        """
+
+        if isinstance(shape, int):
+            d = shape
+            shape = (shape,)
+        else:
+            d = prod(shape)
+
+        if num_threads is None:
+            # only multi-thread for big ``d``
+            if d <= 32768:
[…]
