+++ tests/test_gen/test_rand.py
@@ -1,65 +1,63 @@
 import pytest
-from numpy.testing import assert_allclose, assert_almost_equal
+from numpy.testing import assert_allclose
 import numpy as np
 import scipy.sparse as sp
-from quimb import (
-    isherm,
-    ispos,
-    eye,
-    tr,
-    ptr,
-    expec,
-    eigvalsh,
-    mutual_information,
-    logneg,
-    rand_matrix,
-    rand_herm,
-    rand_pos,
-    rand_rho,
-    rand_ket,
-    rand_uni,
-    rand_haar_state,
-    gen_rand_haar_states,
-    rand_mix,
-    rand_product_state,
-    rand_matrix_product_state,
-    rand_seperable,
-)
+import quimb as qu
 
 
 dtypes = [np.float32, np.float64, np.complex128, np.complex64]
 
 
+class TestRandn:
+
+    @pytest.mark.parametrize(
+        'dtype', dtypes + [float, complex, 'f8', 'f4', 'c8', 'c16', 'float32',
+                           'float64', 'complex64', 'complex128'])
+    def test_basic(self, dtype):
+        x = qu.randn((2, 3, 4), dtype=dtype)
+        assert x.shape == (2, 3, 4)
+        assert x.dtype == np.dtype(dtype)
+
+    def test_can_seed(self):
+        assert_allclose(qu.randn(5, seed=42), qu.randn(5, seed=42))
+
+    def test_scale_and_loc(self):
+        x = qu.randn(1000, scale=100, loc=50, dtype=float, seed=42)
+        assert_allclose(np.mean(x), 50, rtol=1e-1)
+        assert_allclose(np.std(x), 100, rtol=1e-1)
+
+
 @pytest.mark.parametrize('dtype', dtypes)
 class TestRandMatrix:
     def test_rand_matrix(self, dtype):
-        a = rand_matrix(3, scaled=True, dtype=dtype)
+        a = qu.rand_matrix(3, scaled=True, dtype=dtype)
         assert a.shape == (3, 3)
         assert type(a) == np.matrix
         assert a.dtype == dtype
 
     def test_rand_matrix_sparse(self, dtype):
-        a = rand_matrix(3, sparse=True, dtype=dtype)
+        a = qu.rand_matrix(3, sparse=True, dtype=dtype)
         assert a.shape == (3, 3)
         assert type(a) == sp.csr_matrix
         assert a.dtype == dtype
 
     def test_rand_matrix_sparse_density(self, dtype):
-        a = rand_matrix(3, sparse=True, density=1 / 9, dtype=dtype)
+        a = qu.rand_matrix(3, sparse=True, density=1 / 9, dtype=dtype)
         assert a.nnz == 1
-        a = rand_matrix(3, sparse=True, density=7 / 9, dtype=dtype)
+        a = qu.rand_matrix(3, sparse=True, density=7 / 9, dtype=dtype)
         assert a.nnz == 7
 
     def test_rand_matrix_bsr(self, dtype):
-        a = rand_matrix(10, sparse=True, density=0.2, stype='bsr', dtype=dtype)
+        a = qu.rand_matrix(10, sparse=True, density=0.2,
+                           stype='bsr', dtype=dtype)
         assert a.shape == (10, 10)
         assert type(a) == sp.bsr_matrix
         assert a.dtype == dtype
 
     @pytest.mark.parametrize('sparse', [False, True])
     def test_seed(self, dtype, sparse):
-        a = rand_matrix(10, sparse=sparse, dtype=dtype, seed=42)
[…]
