### 2018-08-15 289bafa
updates for new rand changes
