### 2018-08-15 ab2bc5d
updates to docs
