+++ docs/basics.rst
@@ -8,13 +8,13 @@ Basic Representation
 States and operators in :py:mod:`quimb` are simply dense numpy or sparse scipy complex matrices.
 The :py:func:`~quimb.core.quimbify` function (aliased to :py:func:`~quimb.core.qu`) can convert between the various representations.
 
-.. code-block:: python
+.. code-block:: py3
 
     >>> data = [1, 2j, -3]
 
 Kets are column vectors, i.e. with shape ``(d, 1)``:
 
-.. code-block:: python
+.. code-block:: py3
 
     >>> qu(data, qtype='ket')
     matrix([[ 1.+0.j],
@@ -25,14 +25,14 @@ The ``normalized=True`` option can be used to ensure a normalized output.
 
 Bras are row vectors, i.e. with shape ``(1, d)``:
 
-.. code-block:: python
+.. code-block:: py3
 
     >>> qu(data, qtype='bra')  # also conjugates the data
     matrix([[ 1.-0.j,  0.-2.j, -3.-0.j]])
 
 And operators are square matrices, i.e. have shape ``(d, d)``:
 
-.. code-block:: python
+.. code-block:: py3
 
     >>> qu(data, qtype='dop')
     matrix([[ 1.+0.j,  0.-2.j, -3.-0.j],
@@ -41,7 +41,7 @@ And operators are square matrices, i.e. have shape ``(d, d)``:
 
 Which can also be sparse:
 
-.. code-block:: python
+.. code-block:: py3
 
     >>> qu(data, qtype='dop', sparse=True)
     <3x3 sparse matrix of type '<class 'numpy.complex128'>'
@@ -65,7 +65,7 @@ Basic Operations
 
 The 'dagger', or hermitian conjugate, operation is performed with the ``.H`` attribute:
 
-.. code-block:: python
+.. code-block:: py3
 
     >>> psi = 1.0j * bell_state('psi-')
     >>> psi
@@ -104,14 +104,14 @@ There are a number of ways to combine states and operators, i.e. tensoring them 
 
 Functional form using :py:func:`~quimb.accel.kron`:
 
-.. code-block:: python
+.. code-block:: py3
 
     >>> kron(psi1, psi2, psi3, ...)
     ...
 
 This can also be done using the ``&`` overload on numpy and scipy matrices:
 
-.. code-block:: python
+.. code-block:: py3
 
     >>> psi1 & psi2 & psi3
     ...
@@ -122,7 +122,7 @@ This can also be done using the ``&`` overload on numpy and scipy matrices:
 
 Often one wants to sandwich an operator with many identities, :py:func:`~quimb.core.ikron` can be used for this:
 
-.. code-block:: python
+.. code-block:: py3
 
     >>> dims = [2] * 10  # overall space of 10 qubits
     >>> X = qu([[0, 1], [1, 0]])  # pauli-X
@@ -132,7 +132,7 @@ Often one wants to sandwich an operator with many identities, :py:func:`~quimb.c
 
 For more advanced tensor constructions, such as reversing and interleaving identities within operators :py:func:`~quimb.core.pkron` can be used:
 
-.. code-block:: python
+.. code-block:: py3
 
     >>> dims = [2] * 3
     >>> XZ = pauli('X') & pauli('Z')
@@ -147,7 +147,7 @@ Here, the internal dimensions of a state must be supplied as well as the indicie
 
 For example, if we have a random system of 10 qubits (hilbert space of dimension ``2**10``), and we want just the reduced density matrix describing the first and last spins:
 
-.. code-block:: python
+.. code-block:: py3
 
     >>> dims = [2] * 10
     >>> D = prod(dims)

[…]
