### 2018-08-16 #819 @sunqm
@jcmgray

Please add the doi for the references.  Most of them should be available This is one item in the JOSS checklist.
