### 2018-08-16 522512f
explicitly depend on tensor module (#3)
* add opt_einsum as dep to setup and installation docs
* add trans_invar option to MPS_rand_state
* use quimb.tensor for rand_mps
* change rand_mps args to (n, bond_dim, phys_dim=2, ...) <- breaking!
