### 2018-08-16 d4f166f
more robust tensor approx spectral tests
