### 2018-08-17 1d4180f
add MERA example notebook
