+++ docs/api.rst
@@ -35,5 +35,6 @@ Autogenerated summary of the modules, functions and classes in :py:mod:`quimb`.
     quimb.tensor.tensor_dmrg
     quimb.tensor.tensor_tebd
     quimb.tensor.tensor_approx_spectral
+    quimb.tensor.tensor_mera
 
 
