### 2018-08-17 0597eca
update paper.bib with DOIs and other info
