+++ paper/paper.bib
@@ -1,9 +1,10 @@
 @book{numpy,
-  title={A guide to NumPy},
-  author={Oliphant, Travis E},
-  volume={1},
-  year={2006},
-  publisher={Trelgol Publishing USA}
+  title = {Guide to NumPy},
+  year = {2015},
+  isbn = {151730007X, 9781517300074},
+  edition = {2nd},
+  publisher = {CreateSpace Independent Publishing Platform},
+  address = {USA},
 }
 
 @Misc{scipy,
@@ -11,7 +12,7 @@
   title =     {{SciPy}: Open source scientific tools for {Python}},
   year =      {2001--},
   url = "http://www.scipy.org/",
-  note = {[Online; accessed <today>]}
+  note = {[Online; accessed 17 August 2018]}
 }
 
 @article{slepc4py,
@@ -32,20 +33,18 @@
   number={9},
   pages={1108--1115},
   year={2005},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={https://doi.org/10.1016/j.jpdc.2005.03.010},
 }
 
 @misc{opt_einsum,
-  author       = {Daniel Smith and
-                  Johnnie Gray and
-                  Fabian-Robert Stöter and
-                  Robert T. McGibbon and
-                  Nils Werner},
-  title        = {dgasmith/opt_einsum: v2.0.0},
-  month        = may,
-  year         = 2018,
-  doi          = {10.5281/zenodo.1248818},
-  url          = {https://doi.org/10.5281/zenodo.1248818}
+  author={Smith, Daniel GA and Gray, Johnnie},
+  title={opt_einsum - A Python package for optimizing contraction order for einsum-like expressions},
+  journal={Journal of Open Source Software},
+  month={june},
+  year={2018},
+  doi={10.21105/joss.00753},
+  url={https://doi.org/10.21105/joss.00753}
 }
 
 @article{tensor-networks,
@@ -57,7 +56,8 @@
   pages={223001},
   url={http://stacks.iop.org/1751-8121/50/i=22/a=223001},
   year={2017},
-  abstract={The curse of dimensionality associated with the Hilbert space of spin systems provides a significant obstruction to the study of condensed matter systems. Tensor networks have proven an important tool in attempting to overcome this difficulty in both the numerical and analytic regimes. These notes form the basis for a seven lecture course, introducing the basics of a range of common tensor networks and algorithms. In particular, we cover: introductory tensor network notation, applications to quantum information, basic properties of matrix product states, a classification of quantum phases using tensor networks, algorithms for finding matrix product states, basic properties of projected entangled pair states, and multiscale entanglement renormalisation ansatz states. The lectures are intended to be generally accessible, although the relevance of many of the examples may be lost on students without a background in many-body physics/quantum information. For each lecture, several problems are given, with worked solutions in an ancillary file.}
[…]
