### 2018-08-17 #819 @jcmgray
I've now added a DOI for as many of the references as I can (https://github.com/jcmgray/quimb/commit/0597ecadf5730f8f87fcbba07f0c0827db722a89).
