### 2018-08-19 #819 @sunqm
@jcmgray Sorry for the long delay for reviewing. Thanks for updating the reference doi. I noticed that the function documents were generated in the _autosummary folder. Although not required by the review checklist, if possible, I think it may be better to have an entry in the main page or in the developer's notes that points to the _autosummary page.

@jedbrown This is a well written program. I have gone through the checklist and everything looks good to me now. Please let me know what I should do next.
