### 2018-08-19 #819 @jedbrown
Great! Thanks, @sunqm and @jcmgray.

@dalcinl Can you weigh in on the remainder of your checklist from your perspective as slepc4py and mpi4py developer?
