### 2018-08-20 #819 @jcmgray
> @jcmgray Sorry for the long delay for reviewing. Thanks for updating the reference doi. I noticed that the function documents were generated in the _autosummary folder. Although not required by the review checklist, if possible, I think it may be better to have an entry in the main page or in the developer's notes that points to the _autosummary page.

Not at all, thanks very much for the review! Yes the ``_autosummary`` is used to generate the sections [here (index/module index)](https://quimb.readthedocs.io/en/latest/#indices-and-tables) as well as cross references within the docs. I think you are right that it might be nice to explicitly point this out on the main page / dev pages, so might add that soon.
