### 2018-08-21 d7946b1
docs updates [ci skip]
