### 2018-08-22 be2468b
various approx accelerations
