+++ quimb/gen/rand.py
@@ -60,8 +60,8 @@ if (
         # all RNGs inherit state from the first RNG of _get_randomgens
         _get_randomgens(1)[0].seed(seed)
 
-    def randn(shape, dtype=float, scale=1.0, loc=0.0,
-              num_threads=None, seed=None):
+    def randn(shape=(), dtype=float, scale=1.0, loc=0.0,
+              num_threads=None, seed=None, dist='normal'):
         """Fast multithreaded generation of random normally distributed data
         using ``randomgen``.
 
@@ -72,11 +72,15 @@ if (
         dtype : {'complex128', 'float64', 'complex64' 'float32'}, optional
             The data-type of the output array.
         scale : float, optional
-            Scale the random distribution by this amount.
+            The width of the distribution (standard deviation if
+            ``dist='normal'``).
         loc : float, optional
-            Shift the random distribution by this amount.
+            The location of the distribution (lower limit if
+            ``dist='uniform'``).
         num_threads : int, optional
             How many threads to use. If ``None``, decide automatically.
+        dist : {'normal', 'uniform'}
+            Type of random number to generate.
         """
         if seed is not None:
             seed_rand(seed)
@@ -96,12 +100,17 @@ if (
 
         rgs = _get_randomgens(num_threads)
 
+        gen_method = {
+            'normal': 'standard_normal',
+            'uniform': 'random_sample'
+        }[dist]
+
         # sequential generation
         if num_threads <= 1:
 
             def create(d, dtype):
                 out = np.empty(d, dtype)
-                rgs[0].generator.standard_normal(out=out, dtype=dtype)
+                getattr(rgs[0].generator, gen_method)(out=out, dtype=dtype)
                 return out
 
         # threaded generation
@@ -117,7 +126,7 @@ if (
             S = math.ceil(d / num_threads)
 
             def _fill(gen, out, dtype, first, last):
-                gen.standard_normal(out=out[first:last], dtype=dtype)
+                getattr(gen, gen_method)(out=out[first:last], dtype=dtype)
 
             def create(d, dtype):
                 out = np.empty(d, dtype)
@@ -156,7 +165,7 @@ if (
         return out.reshape(shape)
 
     def rand(*args, **kwargs):
-        return _get_randomgens(1)[0].generator.rand(*args, **kwargs)
+        return randn(*args, dist='uniform', **kwargs)
 
     def randint(*args, **kwargs):
         return _get_randomgens(1)[0].generator.randint(*args, **kwargs)
@@ -170,7 +179,8 @@ else:
     def seed_rand(seed):
         np.random.seed(seed)
 
-    def randn(shape, loc=0.0, scale=1.0, dtype=float, seed=None):
+    def randn(shape=(), dtype=float, scale=1.0, loc=0.0,
+              seed=None, dist='normal'):
         """Generate normally distributed random array of certain shape and type.
         Like :func:`numpy.random.randn` but can specify ``dtype``.
 
@@ -180,6 +190,14 @@ else:
             The shape of the array.
         dtype : {float, complex, ...}, optional
             The numpy data type.
+        scale : float, optional
+            The width of the distribution (standard deviation if
+            ``dist='normal'``).
+        loc : float, optional
+            The location of the distribution (lower limit if
+            ``dist='uniform'``).
+        dist : {'normal', 'uniform'}
+            Type of random number to generate.
 
         Returns
         -------
@@ -188,16 +206,20 @@ else:
         if seed is not None:
             seed_rand(seed)
 
+        def create():
+            if dist == 'normal':
[…]
