### 2018-08-22 ded1e29
custom, much faster random sparse matrix construction
