### 2018-08-22 f703f6c
fix tests
make sure sparse random matrix uses unique pairs but efficiently
