### 2018-08-23 38b968f
make tests more robust
