### 2018-08-23 d24dbab
tweak tests again, set tau = tol / 3, set k_max = 512
