### 2018-08-23 f5afc9f
randn crept into base_linalg
