### 2018-08-23 #819 @jcmgray
Just to note that I have added a brief sentence highlighting those sections.
