### 2018-08-24 60e45e7
fix some rand related tests
