### 2018-08-30 efeaa86
tweak docs, simplify numba code
