### 2018-08-30 649e17f
make tests more robust
