+++ tests/test_tensor/test_tensor_dmrg.py
@@ -338,7 +338,7 @@ class TestDMRGX:
         assert_allclose(big_ovlps, [1])
 
         # check fully
-        assert is_eigenvector(k, h, atol=1e-12, rtol=1e-12)
+        assert is_eigenvector(k, h, tol=1e-12)
 
     def test_solve_bigger(self):
         n = 14
