### 2018-08-30 f663f9e
better default lanczos precision in approx_spectral
