### 2018-08-30 98997ec
check ints against numbers.Integral
