### 2018-08-31 a515dd6
.A property for TNLOs to match sparse matrices
