### 2018-09-01 1d58095
TN gate: add propagate_tags='register'
