### 2018-09-02 #819 @dalcinl
@jcmgray Sorry for the long delay. I'm quite happy you are taking advantage of the `mpi4py.futures` framework!. There is just one minor issue with the references in the paper. The citation for `slepc4py` is wrong. There is no published paper about `slepc4py`, the suggested citation would be just the Bitbucket repo https://bitbucket.org/slepc/slepc4py, or maybe this paper https://doi.org/10.1016/j.advwatres.2011.04.013 (though it is about `mpi4py` and `petsc4py`, not `slepc4py`). And maybe you should cite core SLEPc: https://doi.org/10.1145/1089014.1089019 (I think this is the paper you have in the current bib file, although the title and authors are wrong).
