### 2018-09-02 3553555
update paper references
