+++ paper/paper.bib
@@ -15,14 +15,28 @@
   note = {[Online; accessed 17 August 2018]}
 }
 
-@article{slepc4py,
-  title={SLEPc for Python},
-  author={Dalcin, Lisandro},
-  journal={Trans. Math. Softw},
+@article{dalcin2011parallel,
+  title={Parallel distributed computing using Python},
+  author={Dalcin, Lisandro D and Paz, Rodrigo R and Kler, Pablo A and Cosimo, Alejandro},
+  journal={Advances in Water Resources},
+  volume={34},
+  number={9},
+  pages={1124--1139},
+  year={2011},
+  publisher={Elsevier},
+  doi={10.1016/j.advwatres.2011.04.013}
+}
+
+@article{hernandez2005slepc,
+  title={SLEPc: A scalable and flexible toolkit for the solution of eigenvalue problems},
+  author={Hernandez, Vicente and Roman, Jose E and Vidal, Vicente},
+  journal={ACM Transactions on Mathematical Software (TOMS)},
   volume={31},
   number={3},
   pages={351--362},
-  year={2005}
+  year={2005},
+  publisher={ACM},
+  doi={10.1145/1089014.1089019}
 }
 
 @article{mpi4py,
@@ -34,7 +48,7 @@
   pages={1108--1115},
   year={2005},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/j.jpdc.2005.03.010},
+  doi={10.1016/j.jpdc.2005.03.010},
 }
 
 @misc{opt_einsum,
