### 2018-09-03 b61172b
add fast, randomized SVD and rank estimation
tweaks to rand_linalg
TNLO-1D: proper _matmat and _adjoint
finalize rand_linalg
