+++ quimb/linalg/rand_linalg.py
@@ -1,0 +1,373 @@
+"""Randomized iterative methods for decompositions.
+"""
+from numbers import Integral
+
+import numpy as np
+import scipy.linalg as sla
+from cytoolz import identity
+
+from ..gen.rand import randn
+from ..accel import dot, njit
+
+
+def dag(X):
+    try:
+        return X.H
+    except AttributeError:
+        return X.conj().T
+
+
+def lu_orthog(X):
+    return sla.lu(X, permute_l=True, overwrite_a=True, check_finite=False)[0]
+
+
+def qr_orthog(X):
+    return sla.qr(X, mode='economic', overwrite_a=True, check_finite=False)[0]
+
+
+def orthog(X, lu=False):
+    if lu:
+        return lu_orthog(X)
+    return qr_orthog(X)
+
+
+def QB_to_svd(Q, B, compute_uv=True):
+    UsV = sla.svd(B, full_matrices=False, compute_uv=compute_uv,
+                  overwrite_a=True, check_finite=False)
+
+    if not compute_uv:
+        return UsV
+
+    U, s, V = UsV
+    return dot(Q, U), s, V
+
+
+def trim_usv(UsVH, k):
+    if isinstance(UsVH, tuple):
+        U, s, VH = UsVH
+        U, s, VH = U[:, :k], s[:k], VH[:k, :]
+        return U, s, VH
+    else:
+        # just singular values
+        return UsVH[:k]
+
+
+def rsvd_core(A, k, compute_uv=True, q=2, p=0, state=None, AH=None):
+    """Core R3SVD algorithm.
+
+    Parameters
+    ----------
+    A : linear operator, shape (m, n)
+        Operator to decompose, assumed m >= n.
+    k : int
+        Number of singular values to find.
+    compute_uv : bool, optional
+        Return the left and right singular vectors.
+    q : int, optional
+        Number of power iterations.
+    p : int, optional
+        Over sampling factor.
+    state : {None, array_like, (), (G0,), (U0, s0, VH0, G0)}, optional
+        Iterate based on these previous results:
+
+            - None: basic mode.
+            - array_like: use this as the initial subspace.
+            - (): begin block iterations, return G
+            - (G0,) : begin block iterations with this subspace
+            - (U0, s0, VH0, G0): continue block iterations, return G
+
+    """
+    maybe_project_left = maybe_project_right = identity
+
+    # generate first block
+    if state is None or state is ():
+        G = randn((A.shape[1], k + p), dtype=A.dtype)
+
+    # initial block supplied
+    elif hasattr(state, 'shape'):
+        G = state
+    elif len(state) == 1:
+        G, = state
+
+    # mid-way through adaptive algorithm
+    else:
+        U0, s0, VH0, G = state
+        UH0, V0 = dag(U0), dag(VH0)
+
+        def maybe_project_left(X):
+            X -= dot(U0, dot(UH0, X))
+            return X
+
+        def maybe_project_right(X):
+            X -= dot(V0, dot(VH0, X))
+            return X
+
+    iterating = isinstance(state, tuple)
+
+    if AH is None:
+        AH = dag(A)
+
+    kG = G.shape[1]
+    if kG > k + p:
+        G = G[:, :k + p]
+    elif kG < k + p:
+        G_new = randn((A.shape[1], k + p - kG), dtype=A.dtype)
+        G = np.concatenate((G, G_new), axis=1)
+
+    G = maybe_project_right(G)
+    G = orthog(G, lu=True)
+
+    Q = dot(A, G)
+    Q = maybe_project_left(Q)
+    Q = orthog(Q, lu=q > 0)
+
+    # power iterations with stabilization
[…]
