+++ docs/api.rst
@@ -28,6 +28,7 @@ Autogenerated summary of the modules, functions and classes in :py:mod:`quimb`.
     quimb.linalg.slepc_linalg
     quimb.linalg.mpi_launcher
     quimb.linalg.approx_spectral
+    quimb.linalg.rand_linalg
 
     quimb.tensor.tensor_core
     quimb.tensor.tensor_gen

[…]
