### 2018-09-03 851702a
tweak docs
