### 2018-09-03 #819 @jcmgray
@dalcinl thanks very much for getting round to it - ``mpi4py.futures`` has been very useful! Yes wasn't sure about the ``slepc4py`` reference but have now [updated it](https://github.com/jcmgray/quimb/commit/35535551400a62e5c3c8cb8bda969259ee872e51) with those papers you suggest.
