### 2018-09-03 #819 @dalcinl
@jedbrown The couple issues I found are fixed. Everything looks good to me now.
