### 2018-09-04 #819 @jedbrown
@dalcinl Thanks!

@jcmgray I notice one minor grammatical issue in the first paragraph: "requires much care is taken" could be "requires that much care is taken" or "requires great care to be taken". Can you adjust that wording and archive your repository on Zenodo or similar?
