### 2018-09-04 bae8312
update paper.md grammar [ci skip]
