### 2018-09-04 #819 @jcmgray
@jedbrown thanks for picking that up - I've adjusted that wording [here](https://github.com/jcmgray/quimb/commit/bae8312a1320da0721a63aa537f50aa5f4652e70). And I've uploaded the package to zenodo [here](https://zenodo.org/record/1408690).
