### 2018-09-04 #819 @arfon
@dalcinl, @sunqm - many thanks for your reviews here and to @jedbrown for edit this submission ✨

@jcmgray - your paper is now accepted into JOSS and your DOI is https://doi.org/10.21105/joss.00819 :zap: 🚀:boom:
